Fix accents rendering
When the XML files uses accents and special chars (like cedilla "ç"), e.g. the "_DIFF_LANG" string from standard module, when hovering it on Language Management screen, the tooltip text isn't rendered as expected. From what I've seem, it seems to be related on how the translations were stored on XML language on the past, as it replaces accents with their HTML entity counterpart e.g. "é" was stored as "&egrave;" on PHP/XML language files. It seems that when exporting XML files right now, this conversion doesn't happen, so the "_DIFF_LANG" from Portuguese is stored as "Comparação" instead of it's HTML entity. If the XML export is wrong, it should be fixed, but if it's right, this fix make sure that the "Comparação" string is properly rendered on the page.