Typo fixes and fix loading of translations edit screen
When searching for a text on DataTables and reloading the pages, the filtering was still being applied but the text isn't shown on text input. E.g. searching for "calend" on "Diff" field, it filters the results by the "calend" string as expected, but when reloading the page, it was still filtering the results by "calend" but the text isn't shown on "Diff" field. It's a known issue as can be seen here https://stackoverflow.com/questions/49878160/data-table-search-is-empty-after-refresh-but-filter-is-active and would require writing a custom initComplete function to repopulate the search text to the field, but as there's at least three search fields, I couldn't found a way to properly detect which of them has the search str and repopulate it, so I decided to disable the stateSave feature that fixes this behavior. Now every time the page is reloaded the page is loaded from scratch withoutr filtering.