Allows to use Ctrl + A to select all text from translation field and also do not save it's state
When using the Language management it's not possible to use Ctrl + A to select all on input values of the Datatables columns headers filtering feature.
It also disables stateSave because sometimes the headers column filters are empty but they load the state from previous session - it simply filters the content even when the headers column filters of "Language management" page is empty. An ideal solution would gather the previous value of the input and use the stateSave feature but I do not know how to do that.