Allows the SQL upgrade script to run multiple times
In my environment, the installation fails the first time I attempted it, and the original SQL upgrade script do not handle those scenarios so my FormaLMS installation was broken and I had to restore from backup, fix this SQL script and try again. With this fix, it's possible to run this script multiple times without breaking the installation. It checks the database before inserting data or altering the tables structure and only do so if update is needed.