Fix attendance date and dialog images
classroom-dates.html.twig: When adding a new date on a presencial classroom, it was adding the date to DB but as there was a syntax error on JS loop and JSON parsing, it required the user to manually reload the page to see the changes. It also fixes a bug where the date was filled as "24/09/2024" and the table was showing as "23/09/2024" on attendance and only shows the date correctly (24/09/2024) when manually reloading the page. This also fixes a Console JS error about using HTML tags <b> on "option" value (it should use 'striptags' instead of 'raw').
dialog.js: fixes GIF loading display on most FormaLMS dialogs.