Normalize UTF-8 strings on certificates
Most TTF fonts are Unicode instead of UTF-8, e.g. in UTF-8, the character "é" can be represented by "e + '", while Unicode represents it as an entity "é". As most fonts doesn't support UTF-8, they will print "e" followed by an square on the FormaLMS cert because the Font used on TPDF doesn't support UTF-characters, so this fix allows a wider range of fonts be used with TPDF.