Add option "Use email as userid"
This commit adds an checkbox to the system configuration that allows admins to automatically setup FormaLMS to use email address as username instead of allowing the user to specify the username. I noticed that novice users sometimes types it's own name on this field instead of a username to login into the system.