Fix date add and use YUI2 dialog
When adding dates to presencial or e-learning courses editions, it uses a custom modal dialog that seems troublesome when dealing with browser window size (the text isn't centered or shown outside of the modal box), so this PR replaces the current modal system used on this page by the default YUI2 dialog used everywhere else on the system.

This also fixes an issue that when adding a date, it seems to show it on table on wrong date, e.g. adding "05/10/2024" shows "04/10/2024" on table, but added as "05/10/2024" as expected on DB. This fixes the date to show as "05/10/2024" as seen on DB.