Validate start/time and interval hours
Check if a valid value was typed in course begin/end hours and pause begin/end hours.
This fix also removes the "input-not-valid" class if field was validated because when fixing the field value but filling another one with error the fixed field will still be shown as invalid. This PR fixes that.
Also ignore checking if pause begin/end hours is between course begin/end hours if pause hours has been kept with default values (00:00).